orgolife_fixed/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.3
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0
email-validator==2.1.1
slowapi==0.1.9
starlette==0.37.2
httpx==0.27.0
psycopg2-binary==2.9.11
